Refactor(plugins): Move jinja filter code for arista.avd.add_md_toc to PyAVD
