+++ python-avd/pyavd/templater.py
@@ -4,6 +4,7 @@
 from jinja2 import ChoiceLoader, Environment, FileSystemLoader, ModuleLoader, StrictUndefined
 
 from .constants import JINJA2_EXTENSIONS, JINJA2_PRECOMPILED_TEMPLATE_PATH
+from .j2filters.add_md_toc import add_md_toc
 from .j2filters.convert_dicts import convert_dicts
 from .j2filters.default import default
 from .j2filters.natural_sort import natural_sort
@@ -74,6 +75,7 @@ class Templar:
                 "arista.avd.list_compress": list_compress,
                 "arista.avd.natural_sort": natural_sort,
                 "arista.avd.range_expand": range_expand,
+                "arista.avd.add_md_toc": add_md_toc,
             }
         )
         self.environment.tests.update(
