+++ python-avd/tests/pyavd/j2filters/toc_files/invalid_file.md
@@ -4,28 +4,28 @@
   ~ that can be found in the LICENSE file.
   -->
 
+<!---- toc -->
+<!-- toc -->
 # md-toc
+
 a python script to generate the toc (table of contents) of markdown
 
 ## toc
+
 @{md-toc}@
 
 ## todo
+
 - [ ] basic func
 - [ ] unit test
 - [ ] js version and a chrome extension
 
 ## Ussage xx
-### ussage1
-### ussage 2
-## todo2
-# todo3
 
-<avd_unit_test_tag_start>
-- [md-toc](#md-toc)
-  - [toc](#toc)
-  - [todo](#todo)
-  - [Ussage xx](#ussage-xx)
-  - [todo2](#todo2)
-- [todo3](#todo3)
-<avd_unit_test_tag_end>
+### ussage1
+
+### ussage 2
+
+## todo2
+
+# todo3
